chore: add newline at EOF for ResultsViewController
